Hallo?

+++ src/main/java/dk/kea/obopgave/utilities/ICrudRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface ICrudRepository<T> {
     void create(T t);
     List<T> readAll();
-    T read(long id);
+    T read(int id);
     boolean update(T t);
-    boolean delete(long id);
+    boolean delete(int id);
 }
